Relocate Readit Steps
Let's introduce a dedicated file for each Step class.
